# Batch user node state diagnostics and improve layout switching performance