--- a/packages/studio-base/src/players/UserNodePlayer/index.ts
+++ b/packages/studio-base/src/players/UserNodePlayer/index.ts
@@ -13,6 +13,7 @@
 
 import { isEqual, uniq } from "lodash";
 import memoizeWeak from "memoize-weak";
+import ReactDOM from "react-dom";
 import shallowequal from "shallowequal";
 import { v4 as uuidv4 } from "uuid";
 
@@ -748,9 +749,23 @@ export default class UserNodePlayer implements Player {
       this._memoizedNodeDatatypes = nodeDatatypes;
     }
 
-    for (const nodeRegistration of state.nodeRegistrations) {
-      this._setUserNodeDiagnostics(nodeRegistration.nodeId, []);
-    }
+    // We need to set the user node diagnostics, which is a react set state
+    // function. This is called once per user script. Since this is in an async
+    // function, the state updates will not be batched below React 18 and React
+    // will update components synchronously during the set state. In a complex
+    // layout, each of the following _setUserNodeDiagnostics call result in
+    // ~100ms of latency. With many scripts, this can turn into a multi-second
+    // stall during layout switching.
+    //
+    // By batching the state update, unnecessary component updates are avoided
+    // and performance is improved for layout switching and initial loading.
+    //
+    // Moving to React 18 should remove the need for this call.
+    ReactDOM.unstable_batchedUpdates(() => {
+      for (const nodeRegistration of state.nodeRegistrations) {
+        this._setUserNodeDiagnostics(nodeRegistration.nodeId, []);
+      }
+    });
   }
 
   private async _getRosLib(state: ProtectedState): Promise<string> {